refactor: expand astra-pipeline with 8 modules (~7K LOC)
Extract 8 pipeline modules from runtime to astra-pipeline crate:
- pattern.rs (2,433 LOC) — tool chain pattern library with PatternAction type
- task_learning.rs (1,005 LOC) — per-task learning feedback aggregation
- scheduling.rs (880 LOC) — step scheduling contracts
- mo_persistence.rs (731 LOC) — MO cloud checkpoint persistence
- step_restore.rs (663 LOC) — checkpoint restore logic
- stages/reflect.rs (503 LOC) — reflection pipeline stage
- stages/evaluate.rs (456 LOC) — evaluation pipeline stage
- defaults.rs (352 LOC) — pipeline default configurations

PatternAction enum moved from evolution/types.rs to astra-pipeline,
runtime re-exports for backward compatibility.

Added deps: astra-learning, sqlx, tokio, uuid to astra-pipeline.
All 443 astra-pipeline tests pass, runtime pipeline tests pass.
