+++ rust/crates/astra-pipeline/Cargo.toml
@@ -10,7 +10,7 @@ default = []
 
 [dependencies]
 astra-core.workspace = true
-astra-learning = { path = "../astra-learning" }
+astra-learning.workspace = true
 astra-services.workspace = true
 astra-text-utils = { path = "../astra-text-utils" }
 astra-turn-types = { path = "../astra-turn-types" }
