+++ rust/crates/astra-pipeline/src/lib.rs
@@ -9,26 +9,25 @@
 #![allow(clippy::too_many_arguments)]
 
 pub mod calibration;
+pub mod defaults;
 pub mod engine;
 pub mod entity;
 pub mod event;
 pub mod feedback_extraction;
 pub mod feedback_store;
+pub mod mo_persistence;
+pub mod pattern;
 pub mod routing;
+pub mod scheduling;
+pub mod stages;
 pub mod state;
 pub mod step_checkpoint;
 pub mod step_protocol;
 pub mod step_recorder;
-pub mod tool_health_types;
-
-pub mod defaults;
-pub mod evaluate;
-pub mod mo_persistence;
-pub mod pattern;
-pub mod reflect;
-pub mod scheduling;
 pub mod step_restore;
 pub mod task_learning;
+pub mod tool_health_types;
+
 pub use pattern::PatternAction;
 pub use routing::{CalibrationAxis, DomainHint, TaskType, ToolFilter, domain_hint_to_label};
 pub use tool_health_types::ToolHealthEntry;
