+++ rust/crates/runtime/src/pipeline/stages/evaluate.rs
@@ -1,1 +1,1 @@
-pub use astra_pipeline::evaluate::*;
+pub use astra_pipeline::stages::evaluate::*;
