+++ rust/crates/runtime/src/pipeline/stages/reflect.rs
@@ -1,1 +1,1 @@
-pub use astra_pipeline::reflect::*;
+pub use astra_pipeline::stages::reflect::*;
